+++ CHANGELOG.md
@@ -10,6 +10,9 @@ All notable changes to this project should be documented in this file.
     - [Add support for C_[Get|Set]OperationState
        functions](https://github.com/latchset/kryoptic/pull/389)
 
+* Support for CKM_TLS12_MASTER_KEY_DERIVE_DH
+    - [Minor TLS improvements](https://github.com/latchset/kryoptic/pull/399)
+
 --
 
 ## [1.4.0]