+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project should be documented in this file.
 
 ### What Changed
 
+--
+
+## [1.5.0]
+## 2026-03-04
+
+### What Changed
+
 * Added support to save and restore operations state
     - [Add support for C_[Get|Set]OperationState
        functions](https://github.com/latchset/kryoptic/pull/389)
@@ -18,8 +25,6 @@ All notable changes to this project should be documented in this file.
 
 * Refactor code that exposes handles the fips provider and defer KAT self-tests
     - [Fips defer tests](https://github.com/latchset/kryoptic/pull/423)
-
---
 
 ## [1.4.0]
 ## 2025-11-24
@@ -261,3 +266,4 @@ that includes non public interfaces.
 [1.3.0]: https://github.com/latchset/kryoptic/releases/tag/v1.3.0
 [1.3.1]: https://github.com/latchset/kryoptic/releases/tag/v1.3.1
 [1.4.0]: https://github.com/latchset/kryoptic/releases/tag/v1.4.0
+[1.5.0]: https://github.com/latchset/kryoptic/releases/tag/v1.5.0